add assembly files for c# scripts
