--- unity/kuka-ar-unity/Assets/AnchorCreationManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ec6d5cad8075824c877546d166a4849
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
