--- unity/kuka-ar-unity/Assets/AnchorWithChildManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ec9daa57191d264788145998a76301c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
