--- unity/kuka-ar-unity/Assets/ExtendedAnchorCreationManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbe45ff0deec6c2429827054a821d585
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
