+++ unity/kuka-ar-unity/Assets/Project/Scripts/AnchorSystem/AnchorManager.cs.meta
