+++ unity/kuka-ar-unity/Assets/Project/Scripts/AnchorSystem/Scripts.AnchorSystem.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "AnchorSystem",
+    "rootNamespace": "",
+    "references": [
+        "Scripts",
+        "Unity.XR.ARFoundation",
+        "Unity.XR.ARSubsystems"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
