+++ unity/kuka-ar-unity/Assets/Project/Scripts/AnchorSystem/Scripts.AnchorSystem.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eeb86127709eb3b479f586c700f85505
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
