+++ unity/kuka-ar-unity/Assets/Project/Scripts/ImageSystem.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 983d25a8505b6e04b8f657195b2cb5f9
+guid: ccf2b56a31eed39498b51dc533f570af
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
