+++ unity/kuka-ar-unity/Assets/Project/Scripts/ImageSystem/ImageSystem.asmdef
@@ -1,0 +1,19 @@
+{
+    "name": "ImageSystem",
+    "rootNamespace": "",
+    "references": [
+        "GUID:ad47d37b55b3cfa4ba4960d7128606d5",
+        "GUID:a9420e37d7990b54abdef6688edbe313",
+        "GUID:92703082f92b41ba80f0d6912de66115",
+        "GUID:eeb86127709eb3b479f586c700f85505"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
