+++ unity/kuka-ar-unity/Assets/Project/Scripts/ImageSystem/ImageSystem.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: db3c178c407731844af4e0c8fed247d2
-folderAsset: yes
-DefaultImporter:
+guid: bb05188734a25b149898e54d16dd5dd3
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
